% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{array_to_view}
\alias{array_to_view}
\title{array_to_view}
\usage{
array_to_view(arr = arr_img[1, , , ])
}
\arguments{
\item{arr}{an array of dim(h x w x channel) in range 0-255.}
}
\value{
`TRUE`
}
\description{
array_to_view
}
