% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(values, n_sim = 1e+06, p = 0.05)
}
\arguments{
\item{values}{vector of population values.}

\item{n_sim}{number of simulations.}

\item{p}{CI, `0.05 = 95%`.}
}
\value{
a list with mean and CI values.
}
\description{
bootstrap
}
\examples{
\dontrun{
values <- runif(1000, min = 0, max = 1)
bootstrap(values, p = 0.05)
}

}
