% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_to_tex.R
\name{boxplot_to_tex}
\alias{boxplot_to_tex}
\title{df_to_table}
\usage{
boxplot_to_tex(x, x_idx, path, file_name, verbose = FALSE)
}
\arguments{
\item{path}{directory to the .tex file.}

\item{file_name}{name of the .tex file.}

\item{verbose}{default `FALSE`. Set to `TRUE` for printing the string.}

\item{df}{data frame or tibble}

\item{v_lines}{vertical lines in table.}

\item{h_lines}{horizontial lines in table.}

\item{case}{correspond to different types of tables.}

\item{n_digits}{The number of digits in the table. Default is 4 for numeric columns.}

\item{model_name}{Model name for CM.}

\item{norm_cm}{is CM normalized?}
}
\value{
saved file/string with a default tabular table.
}
\description{
df_to_table
}
\examples{
\dontrun{
  library(tidyverse)
  df <- iris \%>\% group_by(Species) \%>\% summarise_all(funs(mean))
  df_to_table(df, path = "~/Downloads/", file_name = "test", v_lines = NULL, h_lines = c(1), case = 1, verbose = FALSE)
}
}
