% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canny_edge.R
\name{canny_edge}
\alias{canny_edge}
\title{canny_edge}
\usage{
canny_edge(grad, angles, TT, tt, strong = 255, weak = 50)
}
\arguments{
\item{grad}{magnitude of the gradients}

\item{angles}{angles of the gradients}

\item{TT}{strong thredshold used in canny edge detection}

\item{tt}{weak thredshold used in canny edge detection}

\item{strong}{strong value}

\item{weak}{weak value}
}
\value{
a binary gray-scale image
}
\description{
edge detection algorithm
}
