% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_weights.R
\name{class_weights}
\alias{class_weights}
\title{class_weights}
\usage{
class_weights(labels, case = 1, mu = 0.15)
}
\arguments{
\item{labels}{vector of labels.}

\item{case}{choose the weight function.}

\item{mu}{scalar to the log case.}
}
\value{
list with weigths to keras
}
\description{
class_weights
}
\examples{
\\donotrun{
labels <- sample(1:7, 100, T)
class_weights(labels)
}

}
