% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_to_tex.R
\name{constants_to_tex}
\alias{constants_to_tex}
\title{constants_to_tex}
\usage{
constants_to_tex(my_list, n_digits = 4L,
  path = "~/MSc/Writing/Report/auto_tbl_fig/",
  file_name = "constants_to_tex", verbose = FALSE)
}
\arguments{
\item{my_list}{list with key and values}

\item{n_digits}{The number of digits in text. Default is 4 for numeric values.}

\item{path}{directory to the .tex file.}

\item{file_name}{name of the .tex file.}

\item{verbose}{default `FALSE`. Set to `TRUE` for printing the string.}
}
\value{
saved file/string with multiple `"\def\key{...}"`
}
\description{
constants_to_tex
}
\examples{
\dontrun{
  #
  my_list <- list("keyone" ="test12456-12.2",
                  "keytwo"="10.1")
}
}
