% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repeated_cv_fold.R
\name{create_repeated_cv_fold}
\alias{create_repeated_cv_fold}
\title{create_repeated_cv_fold}
\usage{
create_repeated_cv_fold(idx, pct_test, n_rep, n_fold, grid)
}
\arguments{
\item{idx}{indecies}

\item{pct_test}{percentage of indecies for test}

\item{n_rep}{number of CV repeats}

\item{n_fold}{number of CV folds}

\item{grid}{search grid}
}
\value{
a expanded list
}
\description{
create_repeated_cv_fold
}
\examples{
\dontrun{
create_repeated_cv_fold(idx = 1:100,
                        pct_test = 0.2,
                        n_rep = 10,
                        n_fold = 5,
                        grid = expand.grid(mtry = 1:2, ntree = 2:5))

}
}
