% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_entropy_loss.R
\name{cross_entropy_loss}
\alias{cross_entropy_loss}
\title{cross_entropy_loss}
\usage{
cross_entropy_loss(predictions, targets, epsilon = 1e-12,
  log_epsilon = 1e-10)
}
\arguments{
\item{predictions}{one-hot encoded proba matrix (with (n,k)).}

\item{targets}{one-hot encoded matrix (with (n,k)).}

\item{epsilon}{clipping offset.}

\item{log_epsilon}{log offset.}
}
\value{
numeric loss.
}
\description{
Computes cross entropy loss.
}
\examples{
\dontrun{
cross_entropy_loss(matrix(c(0.25,0.01,0.25,0.01,0.25,0.01,0.25,0.96), nrow = 2),
                   matrix(c(0,0,0,0,0,0,1,1), nrow=2))
}

}
