% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_gantt.R
\name{df_to_gantt}
\alias{df_to_gantt}
\title{df_to_gantt}
\usage{
df_to_gantt(df, path, file_name, start_date = as.Date("2018-08-01",
  "\%Y-\%m-\%d"), end_date = as.Date("2019-02-01", "\%Y-\%m-\%d"),
  verbose = FALSE)
}
\arguments{
\item{df}{data frame or tibble}

\item{path}{directory to the .tex file.}

\item{file_name}{name of the .tex file.}

\item{start_date}{start date of project,}

\item{end_date}{end date of project.}

\item{verbose}{default `FALSE`. Set to `TRUE` for printing the string.}
}
\value{
saved file/string with a default tabular table.
}
\description{
df_to_gantt
}
\examples{
\dontrun{
  # hmm
}
}
