% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_table.R
\name{df_to_table}
\alias{df_to_table}
\title{df_to_table}
\usage{
df_to_table(df, path, file_name, v_lines = NULL, h_lines = c(1),
  case = 1, n_digits = 4, verbose = FALSE)
}
\arguments{
\item{df}{data frame or tibble}

\item{path}{directory to the .tex file.}

\item{file_name}{name of the .tex file.}

\item{v_lines}{vertical lines in table.}

\item{h_lines}{horizontial lines in table.}

\item{case}{correspond to different types of tables.}

\item{n_digits}{The number of digits in the table. Default is 4 for numeric columns.}

\item{verbose}{default `FALSE`. Set to `TRUE` for printing the string.}
}
\value{
saved file/string with a default tabular table.
}
\description{
df_to_table
}
\examples{
\dontrun{
  library(tidyverse)
  df <- iris \%>\% group_by(Species) \%>\% summarise_all(funs(mean))
  df_to_table(df, path = "~/Downloads/", file_name = "test", v_lines = NULL, h_lines = c(1), case = 1, verbose = FALSE)
}
}
