% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kappa.R
\name{eval_kappa}
\alias{eval_kappa}
\title{eval_kappa}
\usage{
eval_kappa(y_act = sample(1:5, 400, TRUE), y_pred = sample(1:5, 400,
  TRUE), order_classes = c(1, 2, 3, 4, 5))
}
\arguments{
\item{y_act}{true labels}

\item{y_pred}{actual labels}

\item{order_classes}{Sort class order?}
}
\value{
a list with ...
}
\description{
eval_kappa
}
