% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eval_performance}
\alias{eval_performance}
\title{eval_performance}
\usage{
eval_performance(y_act = sample(1:5, 400, TRUE), y_pred = c(sample(1:5,
  200, TRUE), rep(5, 200)), order_classes = c(1, 2, 3, 4, 5),
  prior_classes = NULL, label_classes = c("one", "two", "cat", "cow",
  "dog"), file_name = "test", create_tex_table = TRUE,
  model_name = NULL, norm_cm = TRUE, performace_metric = TRUE)
}
\arguments{
\item{y_act}{correct class}

\item{y_pred}{predicted class}

\item{order_classes}{index of class name}

\item{label_classes}{class labels (unique)}

\item{file_name}{name of tex table file}

\item{create_tex_table}{default is `TRUE`}

\item{model_name}{name of model, default is `NULL`}

\item{norm_cm}{normalize the CM?, default is `TRUE`}

\item{performace_metric}{include performance metrics, default is `TRUE`}
}
\value{
a text table and a list with performance metrics.
}
\description{
eval_performance
}
