% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_train_and_test}
\alias{get_train_and_test}
\title{get_train_and_test}
\usage{
get_train_and_test(df, n_top = 5, shuffle = TRUE, balanced = TRUE,
  pct = 1/5)
}
\arguments{
\item{df}{input tibble.. must be `glo_df`}

\item{n_top}{an integer for groupen the most `n_top` classes and the rest to another signle class.}

\item{shuffle}{shuffle the rows before splitting}

\item{balanced}{balance the number of classes in train and test.}

\item{pct}{percentage to split into test.}
}
\value{
`sagsnr`, `test` and `label`.. must bed left joined to `glo_df`
}
\description{
get_train_and_test
}
