% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOG.R
\name{hog_descriptor}
\alias{hog_descriptor}
\title{hog_descriptor}
\usage{
hog_descriptor(img, cell, n_bins)
}
\arguments{
\item{img}{a gray-scale matrix.}

\item{cell}{the number of HOG windows per bound box.}

\item{n_bins}{the number of histogram bins.}
}
\value{
a feature vector of length `cell^2 * n_bins`.
}
\description{
Histogram of Orientated Gradients (HOG)..
}
