% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hough_transform.R
\name{hough_transform}
\alias{hough_transform}
\title{hough_transform}
\usage{
hough_transform(img, TT, tt, degree, d_rho)
}
\arguments{
\item{img}{a gray_scale image "(,,1)"}

\item{TT}{strong thredshold used in canny edge detection}

\item{tt}{weak thredshold used in canny edge detection}

\item{degree}{vector with possible rotations}

\item{d_rho}{rho resolution}
}
\value{
an image rotation
}
\description{
hough_transform
}
