% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr_package.R
\name{lintr_package}
\alias{lintr_package}
\title{Lintr current package}
\usage{
lintr_package()
}
\description{
Apply one or more linters to all of the R files in a package.
             This function can be used as an addin.
}
