% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_obj}
\alias{load_obj}
\title{Title
This functions creates a new environment and load object to a different "new" object name.}
\usage{
load_obj(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
New opbejct.
}
\description{
Title
This functions creates a new environment and load object to a different "new" object name.
}
