% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{magick_to_array}
\alias{magick_to_array}
\title{magick_to_array}
\usage{
magick_to_array(img, channels = 3)
}
\arguments{
\item{img}{magick obejct.}

\item{channels}{number of channels. fefaults is `3` else a gray-scale image will be returned.}
}
\value{
an array.
}
\description{
This functions converts a magick object to a multidimensionelt integer array.
}
