% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_top}
\alias{my_top}
\title{my_top}
\usage{
my_top(df, split = NULL)
}
\arguments{
\item{df}{input tibble.. must be `glo_df`}

\item{split}{case specific}
}
\value{
an ordered tibble grouped on `country_code` and `country_name`
}
\description{
my_top
}
