% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_days.R
\name{number_of_days}
\alias{number_of_days}
\title{number_of_days}
\usage{
number_of_days(date)
}
\arguments{
\item{date}{as.Date() input}
}
\value{
integer for no. of days.
}
\description{
number_of_days
}
