% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_avg_mean}
\alias{update_avg_mean}
\title{update_avg_mean}
\usage{
update_avg_mean(file_in, file_out = "./data/avg_mean.Rda")
}
\arguments{
\item{file_in}{path to complete array. objects must be named: arr_img}

\item{file_out}{path to save directory. objects must be named: avg_mean}
}
\value{
`TRUE`
}
\description{
This function averages and updates the array (avg_mean) for w.r.t. the given `file_in` array.
This works for 1 and 3 channels
}
