% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GIT_files.R
\name{update_GIT_files}
\alias{update_GIT_files}
\title{update_GIT_files}
\usage{
update_GIT_files(commit_msg, path = "~/MSc/Writing/Report/auto_tbl_fig/",
  files = NULL)
}
\arguments{
\item{commit_msg}{msg which describes the commit.}

\item{path}{directory to the GIT repo.}

\item{files}{files you want to update.}
}
\value{
NULL.
}
\description{
update_GIT_files
}
\examples{
\dontrun{
  update_GIT_files(commit_msg = "test", path = "~/Downloads/test/", files = NULL)
}

}
