% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_to_tikz_line.R
\name{xy_to_tikz_line}
\alias{xy_to_tikz_line}
\title{xy_to_tikz_line}
\usage{
xy_to_tikz_line(x, y, color, legend, path, file_name, round_n = 3,
  line_type = "", verbose = FALSE, fill_opa = 1, sacle = 2)
}
\arguments{
\item{x}{first coordinate.}

\item{y}{second coordinate.}

\item{color}{color of point.}

\item{legend}{name of the realization.}

\item{path}{directory to the .tex file.}

\item{file_name}{name of the .tex file.}

\item{round_n}{number of decimal digits.}

\item{line_type}{default is normal. dashed, dotted is possible.}

\item{verbose}{default `FALSE`. Set to `TRUE` for printing the string.}

\item{fill_opa}{fill opacity of the line.}

\item{sacle}{scaling the size of the line.}
}
\value{
saved file/string with a default tikz picture.
}
\description{
xy_to_tikz_line
}
\examples{
\dontrun{
  library(homemade)
  library(tidyverse)
  df1 <- iris \%>\% filter(Species == "setosa") \%>\% mutate(x = Sepal.Length, y = Petal.Length)
  df2 <- iris \%>\% filter(Species == "versicolor") \%>\% mutate(x = Sepal.Length, y = Petal.Length)

  xy_to_tikz_line(x = df1$x, y = df1$y, color = "red", legend = "setosa", path = "~/Downloads/", file_name = "setosa")
  xy_to_tikz_line(x = df2$x, y = df2$y, color = "blue", legend = "versicolor", path = "~/Downloads/", file_name = "versicolor")
}

}
