% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_to_tikz_point.R
\name{xy_to_tikz_point}
\alias{xy_to_tikz_point}
\title{xy_to_tikz_point}
\usage{
xy_to_tikz_point(x, y, color, legend, path = "~/Downloads/",
  file_name = "setosa", round_n = 3, mark = "*", verbose = FALSE,
  fill_opa = 1, sacle = 1)
}
\arguments{
\item{x}{first coordinate.}

\item{y}{second coordinate.}

\item{color}{color of point.}

\item{legend}{name of the realization.}

\item{path}{directory to the .tex file.}

\item{file_name}{name of the .tex file.}

\item{round_n}{number of decimal digits.}

\item{mark}{marking style.}

\item{verbose}{default `FALSE`. Set to `TRUE` for printing the string.}

\item{fill_opa}{fill opacity of the mark.}

\item{sacle}{scaling the size of the mark.}
}
\value{
saved file/string with a default tikz picture.
}
\description{
xy_to_tikz_point
}
\examples{
\dontrun{
  library(homemade)
  library(tidyverse)
  df1 <- iris \%>\% filter(Species == "setosa") \%>\% mutate(x = Sepal.Length, y = Petal.Length)
  df2 <- iris \%>\% filter(Species == "versicolor") \%>\% mutate(x = Sepal.Length, y = Petal.Length)

  xy_to_tikz_point(x = df1$x, y = df1$y, color = "red", legend = "setosa", path = "~/Downloads/", file_name = "setosa_2")
  xy_to_tikz_point(x = df2$x, y = df2$y, color = "blue", legend = "versicolor", path = "~/Downloads/", file_name = "versicolor_2")
}

}
